rename mod to tx
